requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.1
python-telegram-bot==20.6